requirements.txt
pandas==1.2.4
dataclasses==0.8
Pillow==9.0.0
